requirements/dev.txt
-r base.txt

fabric
ipython[notebook]
django-debug-toolbar
django-extensions

Werkzeug

